+++ runtime/src/lib.rs
@@ -476,7 +476,7 @@ impl pallet_treasury::Trait for Runtime {
 
 parameter_types! {
 	pub const HistoricalPayoutEras: u16 = 7;
-	pub const PayoutSplitThreshold: usize = 1000;
+	pub const PayoutSplitThreshold: u32 = 1000;
 }
 impl crml_staking_rewards::Trait for Runtime {
 	type CurrencyToReward = SpendingAssetCurrency<Self>;
